Fixed issue where import_xml_string still attempting to set the state_set when importing even though this is set in the constructor.
Work creation is now done after the title is obtained so that the title slug can be set to a unique value.

The XML content is now maintained even before we have a verse to append the content to.

Note tags are no longer excluded since that now supported.

Division tags are no longer appended to verses.