Card milestones are now considered chunks.
Added the option (ignore_division_markers) that causes the Perseus importer to ignore division markers.

Removed unnecessary log message indicating that no verse content existed to save.

Fixed problem when the first text node was placed in the wrong location for the first verse. This closes #420.

Division markers are now included in the resulting original content.